finished fillers + pace algo and API

+++ app.py
@@ -62,8 +62,10 @@ def evaluation():
         speechRef = userRef.collection(u'speeches').document(speechID)
 
         # Result values initialize
+        wordCount = 0
         fillers = 0
         fillersDesc = ''
+        fillersPct = 0.0
         pace = 0
         paceDesc = ''
         transcript = ''
@@ -78,11 +80,18 @@ def evaluation():
             transcript = transcribe_gcs(file_url=audioURL, filename=fileName, uid=uid)
         if val_audio == fileName:
             transcript = transcribe_gcs(file_url="", filename=fileName, uid=uid)
+        
 
+        wordCount = count_words(transcript)
         pace = get_pace(transcript, audioURL)
+        paceDesc = get_pace_desc(pace)
         fillers = get_fillers(fileName, audioURL)
-        # Remove file after fillers processing
-        os.remove("filler_" + fileName)
+        fillersPct = get_fillers_pct(fillers,wordCount)
+        fillersDesc = get_fillers_desc(fillersPct)
+
+        # Remove file after processing
+        if os.path.isfile("filler_" + fileName):
+            os.remove("filler_" + fileName)
         os.remove(fileName)
 
         
@@ -93,16 +102,16 @@ def evaluation():
             "fileName": fileName,
             "audioURL": audioURL,
             "script": script,
-            "userRef": userRef,
-            "speechRef": speechRef,
             "pace": pace,
             "paceDesc": paceDesc,
             "fillers": fillers,
-            "fillersDesc": fillersDesc
-
+            "fillersDesc": fillersDesc,
+            "fillersPct": fillersPct,
+            "wordCount": wordCount
         })
+
         # Update firebase parameters
-        speechRef.update({u'fillers': 33, u'fillersDesc': 'Needs Improvement'})
+        speechRef.update({u'fillers': fillers ,u'fillersDesc': fillersDesc, u'fillersPct': fillersPct, u'pace': pace, u'paceDesc': paceDesc, u'wordCount': wordCount})
 
         # print results to console
         my_result = speechRef.get()
